+++ common/src/main/java/com/brandwatch/minibcr/common/domain/Query.java
@@ -15,8 +15,6 @@ public class Query {
         this.definition = definition;
     }
 
-    public Query(){}
-
     public long getId() {
         return id;
     }
