--- common/src/main/java/com/brandwatch/minibcr/common/repository/QueryRowMapper.java
@@ -1,21 +1,0 @@
-package com.brandwatch.minibcr.common.repository;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Component;
-
-import com.brandwatch.minibcr.common.domain.Query;
-
-@Component
-public class QueryRowMapper implements RowMapper {
-
-    @Override
-    public Object mapRow(ResultSet resultSet, int i) throws SQLException {
-        Query query = new Query();
-        query.setId(resultSet.getLong("id"));
-        query.setDefinition(resultSet.getString("definition"));
-        return query;
-    }
-}
