+++ common/src/main/java/com/brandwatch/minibcr/common/repository/mention/MentionRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.brandwatch.minibcr.common.repository;
+package com.brandwatch.minibcr.common.repository.mention;
 
 import java.util.List;
