+++ common/src/main/java/com/brandwatch/minibcr/common/repository/mention/MentionRowMapper.java
@@ -1,4 +1,4 @@
-package com.brandwatch.minibcr.common.repository;
+package com.brandwatch.minibcr.common.repository.mention;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -13,9 +13,9 @@ public class MentionRowMapper implements RowMapper<Mention> {
 
     @Override
     public Mention mapRow(ResultSet resultSet, int i) throws SQLException {
-        Mention mention = new Mention();
-        mention.setId(resultSet.getLong("id"));
-        mention.setText(resultSet.getString("text"));
+        long id = resultSet.getLong("id");
+        String text = resultSet.getString("text");
+        Mention mention = new Mention(id, text);
         return mention;
     }
 }
